[txg] Split types and functionality
